Complete field meta-information in property ordering
Add the following new field flags to property ordering:
 - Enum
 - Repeated
 - Map
 - Message

Enum field flag indicates that the field has type of eiher enum or
repeated enum.

Repeated field flag indicates the field is a list of the repeated
values.

Map field flag indicates that the field is a list of key-value pairs
(QHash in QtProtobuf).

Message field flag indicates that the field is a protobuf message.

This is a pre-requisite for the QAbstractProtobufSerializer API
refactoring.

Task-number: QTBUG-123626
Change-Id: Ieadaa80d1cf19cab6d97977f81f20b339298ea92
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>