Fix the sequence of sending the end-of-stream flag
This ensures that initial headers and the initial request are
sent before any data or end-of-stream packet. The initial
issue was in calling the writesDone before the connection to
the HTTP/2 server was established. So at the point when connection
was established the packet queue looked as following:
[{EOS}, {HEADERS}, {REQUEST}]

This happened because headers and initial request only were added
to queue after the the connection is established. The same could
happen to the followup messages sent by client streaming mechanism.

This solution ensures that initial headers are sent first, and
enqueues the initial request immediately at the moment when call/stream
is called.

Fixes: QTBUG-125369
Change-Id: Ie9f737132c8315d6b857d19b0c18c0b1857b6b17
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>