Reintroduce the ProtobufQuick module
The module now makes the required type registrations for QML modules
that uses the generated protobuf types. It registers the core protobuf
types in QML as QML_FOREIGN. The TransparentWrapper types remain the
value types and objects in JS context. Also add the missing
registrations of the respective list types. The type only is missing
the QByteArrayList, which should be registered as QML built-in.

Task-number: QTBUG-125240
Change-Id: Ia03972295fcb04b5a1b4a8822530a0675da77498
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>