Fix stream cancellation and the deadline timeout
The cancel() implementation was not correct in sending an empty data
frame. We need to send the RST_STREAM frame for a proper cancellation.

The deadline timeout now properly cancels the stream and sends the
DeadlineExceeded status-code back to the client.

Rework the test as the previous functionality got barely tested (and was
incorrect too!).

Pick-to: 6.8
Fixes: QTBUG-125375
Change-Id: Iaa3ba60aac887302ae6264d3bbabbccdf666c805
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>