Rid of the (de)serializeEnumList interface
The methods applied the conversion from QList<EnumType> to
QList<QtProtobuf::int64>. This can be done on the QMetaType level
using the covert function. Add the conversion from/to
QList<QtProtobuf::int64> for the QList<EnumType> and the conversion
from/to QStringList, which should be used in JSON serialization.

Pick-to: 6.8
Task-number: QTBUG-123626
Change-Id: Id0300536a8eda3503540089ba7bc29c1cd249752
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>