Privately introduce QProtobufPropertyOrderBuilder
I wanted to write a Builder class for this for some time now because
protobuf technically supports dynamically retrieving new types using
some URL scheme. But we never needed it, so it wasn't done.
Now that we want to treat a Map entry as what it is - a message with
two fields - we need to be able to set up a "property ordering" object
for it. This is what the builder is for.

Change-Id: I3c8da6be98745c5ba270ca320b2898157a1b32e4
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
(cherry picked from commit af54f630a2f3a7411b91a20de3ceb22e02de027e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>