Fix restoring previous iterator in deserializeObject
We were previously just restoring the iterator in one branch.
This made it really hard to debug unrelated issues that happened to
leave through another branch.

This changes what iterator we are checking for validity before leaving
the function, but the iterator shouldn't have consumed more data than
is available in the buffer, so it should be valid anyway.

Pick-to: 6.8 6.7
Change-Id: Iaa818a54961feacb50a20a1f9770a05308cb2485
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>