QProtobufPropertyOrdering: check index before indexing
It's not a problem in our current code, we don't have any code that is
calculating anything out-of-bounds, but for testing the Builder in an
upcoming patch it's easier to have this check in place.

The Builder patch will also serve as a general test for the layout of
the Ordering class.

Pick-to: 6.8
Change-Id: I9a12397a15d3e5c48e95144b385c0e148c8e8046
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>