Grpc: Don't delete signaling objects inside slots
While running one of the tests it crashed in the deleteHandler function
because the http2stream in the handler was deleted while it was
emitting a signal.

To keep the ASSERT correct we disconnect from the signal after
removing the handler.

Change-Id: Icfc5da25138c6d186a69710c1accb2c7893db5f2
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
(cherry picked from commit a44296e3bd3d04c59f6591c465751e1f1035fa1f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>