qtgrpcgen: remove the lambda unary-call overload
It was observed that this overload for unaries is incomplete.
It is swallowing the QGrpcStatus, which we would need.

Instead of adding the QGrpcStatus to the callback remove it entirely.
This overload doesn't add much value, we should not shoot ourselves in
the foot too early.

Change-Id: I887a10cdde03f2c88354c5fc65db814bfd7c7ff5
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 8017462139f88012dcb8769a89cb21317354a870)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>