QGrpcCallReply: remove 2 arg cb subscribe() and provide constraints
Remove the 2 arg callback version of subscribe(). This overload
doesn't add much value. Users can easily call their own error
handler from the 1 arg version.

Constrain the receiver argument to be a correct context type for
functors. Constrain the FinishCallback with a static_assert to be a
compppatible function.

Update the tests. Remove one testcase from the serverstream test. It has
nothing to do there as it's a unary testcase, which gets already tested.

Task-number: QTBUG-123625
Pick-to: 6.8
Change-Id: Ice62f90a2bfec019ffd5ed0724432f877dd61f34
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>