QGrpcClientBase: make RPC requests SCARY
Don't compile the common ground multiple times for different streams but
rather move it to the implementation.

Change trySerialize() to be non-const so that we can emit the
errorOccurred() there.

Remove the call helper and drag the common prefix in the actual RPC
request.

Fix an invalid test-case that tried to connect on a nullptr reply.
The RPC starting logic can return nullptr when, as in this case, invoked
from a different thread and the testcase didn't respected that.

Task-number: QTBUG-123625
Change-Id: Ib4880dbe6d7b156faecec00be46c2e266c85d11e
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit be679ed4d2ce40cd63c38371277455ea60c6b736)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>