QGrpcClientBasePrivate: harden signal emission for failed thread checks
This patch adds sendErrorOccurredForNonMainThread() to send the signal
with AutoConnection in case of a different thread to avoid the UB when
users used DirectConnection in their code.

The QGrpcStatus returned  in 'checkThread' had no usage outside of the
function so refactor to return a boolean.

Remove the 'inThread' testcase for unary calls as this testcase is
already covered by the 'asyncInThread' case, which we rename to
'nonMainThreadIsInvalid' to have a clear and descriptive name for it.

Pick-to: 6.8 6.7
Change-Id: I190927ecf0b4e4abba4bd7edb6b768c3ed26ed9d
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>