Change the (de)serialize handler interface
Replace the QVariant argument property value with the void pointer one.
We should use the inner QVariant data anyway to avoid unnecessary
copies. This change should also address the hidden detaches of the
QVariant values.

Since the pool of (de)serialize handlers is restricted and always
is binded to the exact type, we allow use void pointer as the type
there.

Task-number: QTBUG-123626
Pick-to: 6.8
Change-Id: I47749b8f78b84e3d627efd270260d6545bce3240
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>