QGrpcSerializationFormat: drag 'Format' enum into the qtgrpc namespace
Create the qgrpcnamespace header and put the QtGrpc namespace in there.
This is where this namespace and its children will live.

Drag the QGrpcSerializationFormat::Format enum in there. This makes it
easier to reuse. The QtGrpc will then be exposed to QML so that its
definitions can be reused.

Task-number: QTBUG-123625
Pick-to: 6.8
Change-Id: I9de79a3a77f706ec871b8c21c93f4cafdcb83c29
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>