Move the Qml sender creation to a generated code
Sender is not used in any of the QML make<OperationType> helpers.
We may create it in the QML client code directly and disallow it's
creating outside.

Make the StreamSender privately constructable as the driven-by change,
in the generated code, to prohibit its creation outside the exact
QmlClient implementation.

Pick-to: 6.8
Task-number: QTBUG-120943
Change-Id: I62b9068a957962deaa3411bb37c5cb043f25b431
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>