Disable QML Protobuf tests in Android
The introduced ProtobufQuick module breaks the assignment operator
for the fields of the protobuf scalar types. It seems the issue is
related to the QML_USING macro which is used to register the protobuf
scalar types. If type is not registers it operates normally and
doesn't throw an error. As workaround the assignment can be done
using explicit Number() constructor.

Task-number: QTBUG-127174
Change-Id: I8a8b0c588d0f083fb6a9d35fb35366a3bc01f1cf
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
(cherry picked from commit 3d6c2827a0a8bbb8faa3c7a5a3416a58a1892d5e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>