Fix the invalid mapping of the HTTP/2 errors to gRPC status codes
The StatusCodeMap was relying onto the QNetworkReply::NetworkError
enum, which is not valid. Since the QGrpcHttp2Channel implementation
switched to the use of the Http2Stream the error codes received from
streams are the pure HTTP/2 error codes.

Reimplement the HTTP/2 codes mapping in QGrpcHttp2Channel. Use the
switch/case statement instead of the map, which should give the same
lookup time, but lesser initialization complexity. Also map the codes
correctly, according to the PROTOCOL-HTTP2.md spec.

Any unknown error code is treated as StatusCode::Internal, according to
HTTP/2 RFC recommentation.

Pick-to: 6.7 6.8
Change-Id: I65694c193fc0eaad6cda3f42f820fa844676f7da
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>