QGrpcOperation: provide operation context as (const) ref
This getter is for internal usage in the QGrpcOperation subclasses and
is never non-null. It was provided as a raw pointer to avoid the sharing
with shared_ptr.

Rename it to context() and also provide a const version of it.

Task-number: QTBUG-123625
Change-Id: I7099ae56974abc0af2d19d81993af3755eb40b1c
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
(cherry picked from commit 2af9646e1d9c901e1b520df98cd5ac8be76ce538)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>