QGrpcCallReply: remove subscribe()
The function was introduced back then, when we had finished() separated
into error and ok state. When we switched to the QGrpcStatus finished()
the usefullness of this 'helper' got less and less, where now it's just
a simple wrapper around connect. Remove it now and add the note about
SingleShotConnection in finished()

This makes QGrpcCallReply useless (we could just use QGrpcOperation) but
we'll keep it for potentially further API that could be added for unary
calls only.

Task-number: QTBUG-125406
Change-Id: I80bafe369f79361dc69580a1c11e13ee3104a304
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 788cf57ef70118cd527011570a5fe67e9bab1c3e)
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>