QtGrpc: polish public headers, sort member by getter/setter pairs
As a final step for 6.8 let's polish the headers. That includes
formatting them and also regrouping some class members to be in
getter/setter pairs instead of having all getter or setter combined.
This makes the header, and the features it provides, much easier to
follow. Also this is the schema that we already use in e.g. grpcquick.

Change-Id: I95f9f1e3c75890ae5ea2002dfe3c6a35abf7baf3
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 2e8bff11d6466bec9d7799641330d0619b9a6f29)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>