Make the protobuf serializer registry private
Change the serializing mechanism that is used to serialize protobuf
repeated and map fields. We used the serializer registry to serialize
and deserialize repeated and map fields, which exposed the registry
as public API and was quite ugly. Meanwhile QMetaType allows to
register the mutable zero-copy converter and access meta-type the
special way we want, using QMetaType::registerMutableView.

This introduces QProtobufRepeatedIterator, special class that allows
accessing and adding new elements to the protobuf repeated and map
fields, that are represented as QList<T> and QHash<K, V> respectively.
Register the mutable view to this new class for all related classes.

QProtobufRepeatedIterator is a wrapper around
QtProtobufPrivate::AbstractRepeatedIterator interface and manages the
allocated AbstractRepeatedIterator implementation.

QtProtobufPrivate::AbstractRepeatedIterator got two implementations for
QList and QHash.

The serializer registry is moved to private API but stil can be and is
used to customize the serialization using the proxies.

[ChangeLog][Protobuf] QtProtobufPrivate::SerializationHandler,
QtProtobufPrivate::registerHandler and related functionality are now
private API.
[ChangeLog][Protobuf] Added QProtobufRepeatedIterator. QVariant that
contain repeated or map protobuf values can be converted to this class
using QVariant::view method.

Pick-to: 6.8
Task-number: QTBUG-123626
Change-Id: I282318e82596839ff436bb33a1bed5f3474488fd
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>