Adjust protobuf constrains across repo code
Unify constraints that we use in public API and move them to a
common header file. Add the missing constraints for protobuf types,
that includes all types that sufficient for protobuf processing. Add
map types constains, which checks if the Key and Value meet the
protobuf requirements.

Use newly introduced constaints across the code.

Pick-to: 6.8
Task-number: QTBUG-123626
Change-Id: I7d3944f0a1374e7616304a22ec7ff148967ef525
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>