Any: Remove the internal typecheck condition from Any::unpack
Use the exported specialization, as the obvious thing for the
unpack function. We cannot avoid the alocation of the T object inside
the unpack function anyway, since it's required to be allocated for the
deserializing, but we at least may ensure we don't have unnecessary
allocations done for the unused 'if constexpr' branch.

Also rearrange the Any::unpack method to ensure NRVO. Reset the
preallocated std::optional<T> instead of returning std::nullopt.

Pick-to: 6.8
Task-number: QTBUG-123627
Change-Id: I618c4792c9ea659ebb488d8543476131177d63a1
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>