QGrpcHttp2Channel: detect operations coming from within socket errors
Detect the situation where processOperation is called from within the
Q*Socket::errorOccurred, as it would be the case for immediate
reconnects.

The socket may not be ready for a immediate reconnect attempt and in
such a case the connectTo* function may return without signaling an
error. This would put the RPC in a dangling state, where its not able to
recover and no finished() signal will be emitted.

Solve this by adding a bool that we set during the
Q*Socket::errorOccurred and check during processOperation() calls and do
the asynchronous for our users. Print a warning to remind them about
this, as it's still preferred that users delay their reconnect attempts
on their own.

Ref: https://doc.qt.io/qt-6/qabstractsocket.html#errorOccurred

Pick-to: 6.9 6.8
Task-number: QTBUG-133937
Change-Id: Ic51f661ad576a24716f5790f5e6e2c96a19acbc4
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>