QGrpcHttp2Channel: Prevent infinite error handling in error state
If the channel is in error state the socket could be non writable which
would result in an error loop if reconnections are steadily happening
since the logic would never reach m_reconnectFunction().

Add a safety check for the state to be Connected to protect against this
condition.

Pick-to: 6.8
Fixes: QTBUG-133937
Change-Id: Ie6c4858623ee8d7cc89a57d9bb14e4735fa64117
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 92520e45b875d46364cdc3231b8ddec88de86444)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>