Remove the explicit setting of the accept-encoding header
QHttp2 implementation will make this favor for us, in general the
header is not mandatory for communication with gRPC server, but can
be used by other members of the network communication. Anyway it's
not the level of the gRPC channel where the header should be defined.

Pick-to: 6.9.0 6.8 6.8.3
Change-Id: I73fe3c780bfa373a724255dd26accf37f8a8ff50
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
(cherry picked from commit 39f2692655cdcb4a75b6f85b3fa78f35c44fd964)
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>