QGrpc{Call,Channel}Options: add 'filterServerMetadata'
This option enables to control the filtering of the received server
metadata. Both options provide it as optional<bool>, so that
QAbstractGrpcChannel implementations should choose a reasonable default.

This has not been provided to the QQml*Options, as there is currently no
available API to access the server metadata.

[ChangeLog][QGrpc{Call,Channel}Options] Added the filterServerMetadata
property.

Task-number: QTBUG-138363
Change-Id: I325a3b4f6e68d63f0828a6deb1a7be883247614b
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>