Http2Handler: improve state handling
Since the Http2Handler is largely self-managed, having a clear and
accurate state representation is crucial.

Previously, some key states were missing—particularly those relevant to
the header phase. This made it harder to ensure correct behavior and
state transitions.

This change introduces the 'Idle' and 'RequestHeadersSent' states to
better reflect the handler’s lifecycle. With this we can enforce more
robust validation and ensure the handler behaves as expected.

Furthermore we change the onDone() handling logic (to only allow it to
be called once) to use a bool. Using state management for this is
impractical. It's not a real state of our Http2Handler. It's leaking out
the internal stream state.

Pick-to: 6.9 6.8
Change-Id: I30b5c935c4fe3aaafb66ed91e25562afca2f8804
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 76b8c2626ab4a0386c778e54a0a28414d7727fe6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>