Doc: Fix incorrect usage of \generatelist
The \generatelist command takes an optional argument
that affect how the members are listed, for example,
`classesbymodule` which creates an annotated list of
the classes in a named C++ module.

`groupsbymodule` is not recognized or documented. QDoc
falls back to generating a list of all members of a
group in this case. While the end result is correct,
this relies on an undocumented feature that is likely
to change and break.

Replace these instances with an explicit \annotatedlist
command that produces the same output.

Task-number: QTBUG-138901
Change-Id: Iffa5720e53177d1b996d49f9a607e2a1fa91bb0a
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit bfc975db586dc3629bd14e9f6ec1d2c085405faf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>