Long Live mockserver!
Our current test strategy relies on having a dedicated server running.
While this tests the most 'accurate' flow of the system, it is really
hard to verify the server behavior due to missing flexibility.

Introduce the mockserver to solve this problem. It wraps the low-level
async gRPC server API and allows for non-blocking processing of
operations. This allows it to run nicely together with the Qt Even Loop,
which is blocking.

Provide the server-handling code by using CallbackTags with Lambdas, or
by subclassing the AbstractRpcTag interface.

Change-Id: I086cbd1d6d8542c717ae036bdf54ba96c55afb58
Fixes: QTBUG-139285
Pick-to: 6.8
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit bc052fbd512e781affafc106a917fb50c8b3a284)
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>